Mer lesbar kode
+ logging
